# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Twitter API",
        "description": "Swagger documentation for laravel REST API project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Search for posts and users",
                "description": "Perform a search for posts and users based on a query string.",
                "operationId": "showSearch",
                "requestBody": {
                    "description": "Search query",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "q": {
                                        "type": "string",
                                        "example": "search_query"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "query": {
                                            "type": "string",
                                            "example": "search_query"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tags/trending": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Get trending tags",
                "description": "Retrieve a list of trending tags based on their usage.",
                "operationId": "showTrendingTags",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tags": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "content": {
                                                        "type": "string",
                                                        "example": "tag1"
                                                    },
                                                    "tag_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "string",
                                            "example": "Server error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Bearer Token",
                "scheme": "bearer"
            }
        }
    },
...
}